`!dir$ if()` parser testing commit
able to let the compiler parse `!dir$ if()`, e.g.
```
program mn
!dir$ if(enable)
print *, "INTEL"
!dir$ endif
end
```

However, as it only accept many(letter), and it doesnot do the condition checking.
Compiler the above code will result in:
./test.f90:3:2: warning: Compiler directive was ignored
  !dir$ if(enable)
   ^^^^^^^^^^^^^^^
./test.f90:5:2: warning: Compiler directive was ignored
  !dir$ endif
   ^^^^^^^^^^

Another change can be made is change `!dir$ if()` to `!dir# if()` which will
be parsed as an `#if()`.